Moved images/models into /common/ and added a precache_context_t struct that can be used by both client and server.
Currently it is emplaced and active in the client only. Several MOD_ functions have been modified in order to accept a pointer to a specific content. This has not been done for IMG_ functions, since we currently have no need to load those server side. We merely tag them along in the precache struct for consistency.